+++ src/main/java/com/evacipated/cardcrawl/mod/stslib/powers/interfaces/BetterOnExhaustPower.java
@@ -1,0 +1,9 @@
+package com.evacipated.cardcrawl.mod.stslib.powers.interfaces;
+
+import com.megacrit.cardcrawl.cards.AbstractCard;
+import com.megacrit.cardcrawl.cards.CardGroup;
+
+public interface BetterOnExhaustPower
+{
+    void betterOnExhaust(CardGroup group, AbstractCard card);
+}