+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Changelog ##
 #### dev ####
+* Fix `OnLoseBlock` hooks triggering when creature has no block
 
 #### v1.13.2 ####
 * Fix fetching from exhaust pile making cards invisible