+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 #### dev ####
 * Fix `OnLoseBlock` hooks triggering when creature has no block
 * Allow card selection in `MoveCardsAction` to be sorted
+* Use English strings as backup for localization
 
 #### v1.13.2 ####
 * Fix fetching from exhaust pile making cards invisible