+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fix `OnLoseBlock` hooks triggering when creature has no block
 * Allow card selection in `MoveCardsAction` to be sorted
 * Use English strings as backup for localization
+* Allow `OnReceivePower` to change stackAmount
 
 #### v1.13.2 ####
 * Fix fetching from exhaust pile making cards invisible