+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Allow card selection in `MoveCardsAction` to be sorted
 * Use English strings as backup for localization
 * Allow `OnReceivePower` to change stackAmount
+* Allow `BetterOnApplyPowerPower` to change stackAmount
 
 #### v1.13.2 ####
 * Fix fetching from exhaust pile making cards invisible